% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellClassifier.R
\name{classify.cells}
\alias{classify.cells}
\title{Classify New Embedded Cells}
\usage{
classify.cells(
  data,
  classes,
  k = 7,
  seed = 18051982,
  knn_method = "euclidean",
  knn_weights_fun = NULL,
  nt = 0
)
}
\arguments{
\item{data}{list; GFICF object}

\item{classes}{chareachters; Classes of already existing cells in the order of they are in colnames(data$gficf).}

\item{k}{integer; Number of K-nn to use for classification. Odd number less than 30 are preferred.}

\item{seed}{integer; Initial seed to use.}

\item{knn_method}{; there are various ways of specifying the kernel function. NULL value (default) correspond to unweighted KNN algorithm. See the details section of KernelKnn package for more values.}

\item{nt}{numeric; Number of thread to use (default is 0, i.e. all possible CPUs - 1)}
}
\value{
A dataframe containing cell id and predicted classes.
}
\description{
Classify new embedded cells using GF-ICF transformation and K-nn algorithm.
Existing cells are used as training set.
}
