% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustCells.R
\name{clustcellsBYscGSEA}
\alias{clustcellsBYscGSEA}
\title{Cell clustering by pathway's activity}
\usage{
clustcellsBYscGSEA(
  data,
  pca = NULL,
  method = "phenograph",
  k = 15,
  resolution = 0.25,
  n.start = 50,
  n.iter = 250,
  nt = 0,
  verbose = T,
  seed = 180582
)
}
\arguments{
\item{data}{list; Input data (gficf object)}

\item{pca}{numeric; If different from NULL data are reduced using pca component before to apply clustering algorithm (highly suggested when phenograph methos is choosen).}

\item{method}{character; Clustering method to use such as phenograph or hierarchical.}

\item{k}{integer; number of nearest neighbours (default:15) (used only used if method is phenograph)}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities. (used only used if method is phenograph)}

\item{n.start}{Number of random starts (used only used if method is phenograph).}

\item{n.iter}{Maximal number of iterations per random start (used only used if method is phenograph).}

\item{nt}{integer; Number of cpus to use for k-nn search. If zero all cpu are used.}

\item{verbose}{logical; Increase verbosity.}

\item{seed}{integer; Seed to use for replication.}
}
\value{
the updated gficf object
}
\description{
Cell clustering using pathway expression instead of gene expression
}
\details{
Cells are clustered using the estimated pathway activity levels by scGSEA() function.
Cells are clustered either by using PhenoGraph algorithm or hierarchical clustering. In the case of PhenoGraph method is used
identified clusters with louvain method are stored into data$embedded$cluster.by.scGSEA while the prduced graph into data$scgsea$cell.graph. 
In case hierarchical clusteringis used the resulting dendogram is stored into data$scgsea$h.dendo.
}
