% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensinalityReduction.R
\name{computePCADim}
\alias{computePCADim}
\title{Number of features to use}
\usage{
computePCADim(data, seed = 180582, subsampling = F, plot = T)
}
\arguments{
\item{data}{list; GFICF object}

\item{seed}{numeric; seed to use.}

\item{subsampling}{logical; Use only a subset of the data for the imputation of dimensions to use.}

\item{plot}{logical; Show eblow plot.}
}
\description{
Compute the number of dimension to use for either PCA or LSA.
}
