% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensinalityReduction.R
\name{runNMF}
\alias{runNMF}
\title{Non-Negative Matrix Factorization (NMF)}
\usage{
runNMF(
  data,
  dim = NULL,
  seed = 180582,
  use.odgenes = F,
  n.odgenes = NULL,
  plot.odgenes = F,
  ...
)
}
\arguments{
\item{data}{list; GFICF object}

\item{dim}{integer; Number of dimension which to reduce the dataset.}

\item{seed}{integer; Initial seed to use.}

\item{use.odgenes}{boolean; Use only significant overdispersed genes respect to ICF values.}

\item{n.odgenes}{integer; Number of overdispersed genes to use. A good choise seems to be usually between 1000 and 3000.}

\item{plot.odgenes}{boolean; Show significant overdispersed genes respect to ICF values.}

\item{...}{Additional arguments to pass to nfm call (see ?RcppML::nmf).}

\item{centre}{logical; Centre gficf scores before applying reduction (increase separation).}

\item{randomized}{logical; Use randomized (faster) version for matrix decomposition (default is TRUE).}
}
\value{
The updated gficf object.
}
\description{
Reduce dimensionality of the single cell dataset using Non-Negative Matrix Factorization (NMF)
}
