% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensinalityReduction.R
\name{runPCA}
\alias{runPCA}
\title{Principal Component Analysis (PCA)}
\usage{
runPCA(
  data,
  dim = NULL,
  var.scale = F,
  centre = F,
  seed = 180582,
  use.odgenes = F,
  n.odgenes = NULL,
  plot.odgenes = F
)
}
\arguments{
\item{data}{list; GFICF object}

\item{dim}{integer; Number of dimension which to reduce the dataset.}

\item{centre}{logical; Centre gficf scores before applying reduction (increase separation).}

\item{seed}{integer; Initial seed to use.}

\item{use.odgenes}{boolean; Use only significant overdispersed genes respect to ICF values.}

\item{n.odgenes}{integer; Number of overdispersed genes to use. A good choise seems to be usually between 1000 and 3000.}

\item{plot.odgenes}{boolean; Show significant overdispersed genes respect to ICF values.}
}
\value{
The updated gficf object.

The updated gficf object.
}
\description{
Reduce dimensionality of the single cell dataset using Principal Component Analysis (PCA)
}
