% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensinalityReduction.R
\name{runReduction}
\alias{runReduction}
\title{Dimensionality reduction}
\usage{
runReduction(
  data,
  reduction = "tumap",
  nt = 2,
  seed = 18051982,
  verbose = T,
  ...
)
}
\arguments{
\item{data}{list; GFICF object}

\item{reduction}{characters; Reduction method to use. One of:
\itemize{
  \item \code{"tsne"}
  \item \code{"umap"}
  \item \code{"tumap"} (the default)
}}

\item{nt}{integer; Number of thread to use (default 2).}

\item{seed}{integer; Initial seed to use.}

\item{verbose}{boolean; Icrease verbosity.}

\item{...}{Additional arguments to pass to Rtsne/umap/tumap call.}
}
\value{
The updated gficf object.
}
\description{
Run t-SNE or UMAP or t-UMAP dimensionality reduction on selected features from PCA or NMF.
See ?umap or ?Rtsne for additional parameter to use.
}
