% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_BC_atlas}
\alias{small_BC_atlas}
\title{Single cell breast cancer atlas}
\format{
## `small_BC_atlas`
A Matrix of 4,760 cells
}
\source{
<https://www.nature.com/articles/s41467-022-29358-6>
}
\usage{
small_BC_atlas
}
\description{
A subset of data from the the Single cell breast cancer atlas
Only 150 cells per cell-line are used
}
\keyword{datasets}
