% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{symbolToEns}
\alias{symbolToEns}
\title{Convert Official Gene Symbols to Ensamble IDs}
\usage{
symbolToEns(df, col, organism, verbose = T)
}
\arguments{
\item{df}{data frame; Data frame containing the IDs to convert.}

\item{col}{characters; Name of column containing the gene symbol.}

\item{organism}{characters; Organism of origin (i.e. human or mouse).}

\item{verbose}{boolean; Icrease verbosity.}
}
\value{
The updated data frame with a new column called ens.
}
\description{
It uses biomart. If more the one gene is associated to the enamble, the first one retrived from
Biomart is used.
}
