% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_BC_atlas}
\alias{test_BC_atlas}
\title{Test set for scMAP method}
\format{
## `test_BC_atlas`
A Matrix of 930 cells
}
\source{
<https://www.nature.com/articles/s41467-022-29358-6>
}
\usage{
test_BC_atlas
}
\description{
A subset of data from the the Single cell breast cancer atlas
Only 30 cells per cell-line are used. They do not ovrlapp with cells into small_BC_atlas dataset
}
\keyword{datasets}
