% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalisys.R
\name{runScGSEA}
\alias{runScGSEA}
\title{Single cell Gene Set Enrichement Analysis on GF-ICF}
\usage{
runScGSEA(
  data,
  geneID,
  species,
  category,
  subcategory = NULL,
  pathway.list = NULL,
  nsim = 10000,
  nt = 0,
  minSize = 15,
  maxSize = Inf,
  verbose = TRUE,
  seed = 180582,
  nmf.k = 100,
  fdr.th = 0.05,
  gp = 0,
  rescale = "none"
)
}
\arguments{
\item{data}{list; GFICF object}

\item{geneID}{characters; The type of gene identifier to use, such as ensamble of symbol.}

\item{species}{characters; Species name, such as human or mouse.}

\item{category}{characters; MSigDB collection abbreviation, such as H or C1.}

\item{subcategory}{characters; MSigDB sub-collection abbreviation, such as CGP or BP.}

\item{pathway.list}{list; Custom list of pathways. Each element correspond to a pathway a and contains a vector of genes.}

\item{nsim}{integer; number of simulation used to compute ES significance.}

\item{nt}{numeric; Number of cpu to use for the GSEA and NMF. Default is 0 (i.e., all available cores minus one)}

\item{minSize}{numeric; Minimal size of a gene set to test (default 15). All pathways below the threshold are excluded.}

\item{maxSize}{numeric; Maximal size of a gene set to test (default Inf). All pathways above the threshold are excluded.}

\item{verbose}{boolean; Show the progress bar.}

\item{seed}{integer; Seed to use for random number generation.}

\item{nmf.k}{numeric; Rank of NMF.}

\item{fdr.th}{numeric; FDR threshold for GSEA.}

\item{rescale}{string; If different by none, pathway's activity scores are resealed as Z-score. Possible values are none, byGS or byCell. Default is none.}
}
\value{
The updated gficf object.
}
\description{
Compute GSEA for each cells across a set of input pathways by using NMF.
}
